Updated kafka consumer/producers to use new broker class that allows for compatability with event hubs.
