+++ Notification/Application/Factory/KafkaConsumerFactory.cs
@@ -2,7 +2,6 @@
 using LantanaGroup.Link.Notification.Application.Interfaces;
 using LantanaGroup.Link.Notification.Application.Models;
 using LantanaGroup.Link.Shared.Application.Models;
-using LantanaGroup.Link.Shared.Application.Models.Configs;
 using LantanaGroup.Link.Shared.Application.SerDes;
 using Microsoft.Extensions.Options;
 using System.Diagnostics;
@@ -13,19 +12,21 @@ namespace LantanaGroup.Link.Notification.Application.Factory
     public class KafkaConsumerFactory : IKafkaConsumerFactory
     {
         private readonly ILogger<KafkaConsumerFactory> _logger;
-        private readonly IOptions<KafkaConnection> _kafkaConnection;
+        private readonly IOptions<BrokerConnection> _brokerConnection;
 
-        public KafkaConsumerFactory(ILogger<KafkaConsumerFactory> logger, IOptions<KafkaConnection> kafkaConnection)
+        public KafkaConsumerFactory(ILogger<KafkaConsumerFactory> logger, IOptions<BrokerConnection> brokerConnection)
         {
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-            _kafkaConnection = kafkaConnection ?? throw new ArgumentNullException(nameof(kafkaConnection));
+            _brokerConnection = brokerConnection ?? throw new ArgumentNullException(nameof(brokerConnection));
         }
 
-        public IConsumer<string, NotificationMessage> CreateNotificationRequestedConsumer()
+        public IConsumer<string, NotificationMessage> CreateNotificationRequestedConsumer(bool enableAutoCommit = false)
         {
             try
             {
-                return new ConsumerBuilder<string, NotificationMessage>(_kafkaConnection.Value.CreateConsumerConfig()).SetValueDeserializer(new JsonWithFhirMessageDeserializer<NotificationMessage>()).Build();
+                var config = _brokerConnection.Value.CreateConsumerConfig();
+                config.EnableAutoCommit = enableAutoCommit;
+                return new ConsumerBuilder<string, NotificationMessage>(config).SetValueDeserializer(new JsonWithFhirMessageDeserializer<NotificationMessage>()).Build();
             }
             catch (Exception ex)
             {
