+++ Notification/Application/Interfaces/IKafkaProducerFactory.cs
@@ -5,6 +5,6 @@ namespace LantanaGroup.Link.Notification.Application.Interfaces
 {
     public interface IKafkaProducerFactory
     {
-        public IProducer<string, AuditEventMessage> CreateAuditEventProducer();
+        public IProducer<string, AuditEventMessage> CreateAuditEventProducer(bool useOpenTelemetry);
     }
 }
